feat: migrate fetching data logic to each individual component
